Introduce scala-2 directory to reduce code duplication
