+++ munit/src/main/scala/munit/Location.scala
@@ -3,9 +3,9 @@ package munit
 import java.lang.annotation.Annotation
 import java.io.File
 import scala.runtime.Statics
-import munit.internal.Compat
+import munit.internal.MacroCompat
 
-object Location extends Compat.LocationMacro {
+object Location extends MacroCompat.LocationMacro {
   def empty: Location = new Location("", 0)
 }
